refactor(registrar): extract cert manager
