+++ crates/proof/tee/registrar/src/lib.rs
@@ -10,6 +10,9 @@ pub use config::{
 mod cert_revoker;
 pub use cert_revoker::CertRevoker;
 
+mod cert_manager;
+pub use cert_manager::{CertManager, OnchainRevocationCheck};
+
 mod crl;
 pub use crl::{
     CertCrlInfo, CrlError, DEFAULT_CRL_FETCH_TIMEOUT_SECS, RevokedCertInfo, build_crl_http_client,
@@ -23,8 +26,7 @@ mod driver;
 pub use driver::{
     DEFAULT_MAX_CONCURRENCY, DEFAULT_MAX_TX_RETRIES, DEFAULT_TX_RETRY_DELAY_SECS,
     DEFAULT_UNHEALTHY_REGISTRATION_WINDOW_SECS, DiscoveryResolution, DriverConfig,
-    OnchainRevocationCheck, PendingRegistration, RegisterableSigner, RegistrationDriver,
-    ResolveOutcome,
+    PendingRegistration, RegisterableSigner, RegistrationDriver, ResolveOutcome,
 };
 
 mod error;
