+++ crates/proof/tee/registrar/src/metrics.rs
@@ -25,13 +25,13 @@ base_metrics::define_metrics! {
     #[describe("Total number of certificate revocations detected via CRL")]
     crl_revocations_detected: counter,
 
-    #[describe("Total number of on-chain durable revocation pre-checks performed")]
+    #[describe("Total number of onchain durable revocation pre-checks performed")]
     onchain_revocation_checks_total: counter,
 
-    #[describe("Total number of intermediates rejected by the on-chain durable revocation sentinel")]
+    #[describe("Total number of intermediates rejected by the onchain durable revocation sentinel")]
     onchain_revocations_detected: counter,
 
-    #[describe("Total number of on-chain revocation pre-checks that failed and fell through to the AWS CRL layer (fail-open)")]
+    #[describe("Total number of onchain revocation pre-checks that failed and fell through to the AWS CRL layer (fail-open)")]
     onchain_revocation_check_errors: counter,
 
     #[describe("Total number of revokeCert transaction submission failures")]
